+++ packages/quirk/lib/quirk.dart
@@ -1,7 +1,4 @@
 /// Common utilities, patterns, and extensions for Dart and Flutter.
 library;
 
-export 'src/errors.dart';
-export 'src/iterable_extension.dart';
-export 'src/map_extension.dart';
-export 'src/string_extension.dart';
+export 'src/collection.dart';