+++ packages/quirk/lib/src/collection.dart
@@ -1,0 +1,11 @@
+import 'dart:collection';
+import 'dart:math';
+
+part 'collection/iterable/_delegating_impl.dart';
+part 'collection/iterable/_delegating_mixin.dart';
+part 'collection/list/_delegating_impl.dart';
+part 'collection/list/_delegating_mixin.dart';
+part 'collection/map/_delegating_impl.dart';
+part 'collection/map/_delegating_mixin.dart';
+part 'collection/set/_delegating_impl.dart';
+part 'collection/set/_delegating_mixin.dart';