+++ packages/quirk/test/src/collection/delegating_iterable_test.dart
@@ -1,0 +1,7 @@
+import 'package:quirk/quirk.dart';
+
+import '../suite/test_iterable.dart';
+
+void main() {
+  testIterable(DelegatingIterable.new);
+}