+++ packages/quirk/test/src/collection/delegating_map_test.dart
@@ -1,0 +1,7 @@
+import 'package:quirk/quirk.dart';
+
+import '../suite/test_map.dart';
+
+void main() {
+  testMap(DelegatingMap.new);
+}